requirements.txt
huaweicloudsdkcore
huaweicloudsdkecs
huaweicloudsdkcdn
huaweicloudsdkims
huaweicloudsdkvpc
huaweicloudsdkas
huaweicloudsdkobs
huaweicloudsdkevs
huaweicloudsdkscm
huaweicloudsdkcodeartsbuild
huaweicloudsdkcodeartspipeline
esdk-obs-python
pymongo